New folder structure and some refactoring
- Child coordinators on app coordinator
- Moving logic to the correct classes
