+++ CurrentNewsFeed/Shared/DataManager.swift
