+++ CurrentNewsFeed/Shared/Protocols/Coordinator.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 protocol Coordinator {
-    var childCoordinators: [Coordinator] { get set }
+    var childCoordinators: [Coordinator]? { get set }
     var rootViewController: UIViewController { get }
     
     func start()
